requirements.txt
python-dotenv==1.1.0
selenium==4.33.0
